requirements.txt
accelerate==0.27.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.9.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.42.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.2.2
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.18.0
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
effdet==0.4.1
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
faiss-cpu==1.8.0
fastapi==0.110.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.23.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kor==1.0.1
kubernetes==33.1.0
langchain==0.3.0
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.5
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langgraph==0.2.32
langgraph-checkpoint==2.1.1
langserve==0.3.1
langsmith==0.1.125
layoutparser==0.3.4
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
openai==1.46.1
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.1
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pikepdf==9.10.2
pillow==10.4.0
pillow_heif==0.22.0
portalocker==3.2.0
posthog==6.3.3
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pulsar-client==3.8.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot-ng==2.0.0
Pygments==2.19.2
pyparsing==3.2.3
pypdf==4.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==2.5.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==2.0.0
starlette==0.36.3
streamlit==1.37.0
streamlit-chat==0.1.1
sympy==1.14.0
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.6.0
tiktoken==0.7.0
timm==1.0.19
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.66.2
transformers==4.41.1
triton==3.3.1
trl==0.7.11
typeguard==4.4.4
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.14.1
tyro==0.9.27
unstructured==0.14.2
unstructured-client==0.28.1
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
youtube-transcript-api==0.6.2
zipp==3.23.0
